// projects/console/src/app/interfaces/user.ts
export interface User {
  stripeId?: string;
  stripeLink?: string;
}

export interface OandaCredentials {
  apiKey?: string;
  accountId?: string;
}

export interface Bot {
  status: BotStatus;
  notifications: {
    onStart: boolean;
    onStop: boolean;
    onPing: boolean;
  };
}

export type BotStatus = 'pending' | 'running' | 'stopped';

export type BotCommand = 'ping' | 'stop' | 'start';

export interface Settings {
  bot: {
    status: BotStatus;
    notifications: BotNotificationSettings;
  };
  oanda: {
    credentials: OandaCredentials;
    notifications: OandaNotificationSettings;
  };
  vultr: {
    notifications: VultrNotificationSettings;
  };
}

interface VultrNotificationSettings {
  onStart: boolean;
  onStop: boolean;
}

interface BotNotificationSettings {
  onStart: boolean;
  onStop: boolean;
  onPing: boolean;
}

interface OandaNotificationSettings {
  orders: {
    ordersCancelled: boolean;
    ordersCreated: boolean;
    ordersFilled: boolean;
    ordersTriggered: boolean;
  };
  trades: {
    tradesClosed: boolean;
    tradesOpened: boolean;
    tradesReduced: boolean;
  };
}